fix: tags on discord + assign from discord

+++ api/migrations/17-table-work-on-task.sql
@@ -1,6 +1,6 @@
 CREATE TABLE ctfnote.work_on_task (
     task_id int NOT NULL REFERENCES ctfnote.task (id) ON DELETE CASCADE,
-    profile_id int NOT NULL REFERENCES ctfnote.profile (id) ON DELETE CASCADE,
+    profile_id bigint NOT NULL REFERENCES ctfnote.profile (id) ON DELETE CASCADE,
     PRIMARY KEY (task_id, profile_id)
 );
